requirements.txt
# Hugging Face Spaces (and full local demo): pip install -r requirements.txt
# Lighter local UI only: pip install -r requirements-demo.txt
gradio>=4.44.0,<6
huggingface_hub>=0.23.0
librosa>=0.10.0
matplotlib>=3.8.0
numpy>=1.26.0,<3
python-dotenv>=1.0.0
scipy>=1.11.0
soundfile>=0.12.0
tensorflow-cpu>=2.15.0,<2.19
